slack-webhook/requirements.txt
flask==3.0.0
functions-framework==3.5.0
google-cloud-secret-manager==2.18.2
google-cloud-aiplatform==1.38.1
slack-sdk==3.26.1
